+++ backend/src/main/kotlin/org/aleks616/shrendar/band/model/ArtistBandsDto.kt
@@ -7,7 +7,7 @@ import java.io.Serializable
  */
 data class ArtistBandsDto(
     val id:Int?=null,
-    val artistId:Int?=null,
+    val artistId:Long?=null,
     val artistName:String?=null,
     val bandId:Int?=null,
     val bandName:String?=null,
