+++ backend/src/main/kotlin/org/aleks616/shrendar/contribution/model/ContributionDto.kt
@@ -6,13 +6,13 @@ import java.io.Serializable
  * DTO for {@link org.aleks616.shrendar.contribution.model.Contribution}
  */
 data class ContributionDto(
-    val id:Int?=null,
-    val changeId:Int?=null,
+    val id:Long?=null,
+    val changeId:Long?=null,
     val userId:Int?=null,
     val action:Action?=null,
     val changedTable:String?=null,
     val changedColumn:String?=null,
-    val changedRecordId:Int?=null,
+    val changedRecordId:Long?=null,
     val oldValue:String?=null,
     val newValue:String?=null,
     val changedAt:String?=null,
